requirements.txt
numpy==1.21.4
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu111
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cu111
ipykernel==6.6.0
opencv-python==4.5.4.60
pyyaml==6.0
ipywidgets==7.6.5
matplotlib==3.5.1
tensorboard==2.8.0
jupyter==1.0.0